Move user.py to db (#203)
