+++ src/openlifu/db/__init__.py
@@ -1,9 +1,11 @@
 from .database import Database
 from .session import Session
 from .subject import Subject
+from .user import User
 
 __all__ = [
     "Subject",
     "Session",
-    "Database"
+    "Database",
+    "User",
 ]
